Working on netcdf.py
