+++ test/test_netcdf2d_func.py
@@ -1,0 +1,6 @@
+import pytest
+from s3netcdf.netcdf2d_func import createNetCDF,\
+  writeMetadata,createVariables,createGroupPartition,writeVariable,\
+  getChildShape,getMasterShape
+  
+import numpy as np
